Fix GLesmos flash
